FIxed #1

+++ sdk/common/library/libcpp-spike/include/ColorSensor.h
@@ -23,9 +23,9 @@ class ColorSensor
 public:
 
   struct RGB {
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
+    uint16_t r;
+    uint16_t g;
+    uint16_t b;
   };
 
   struct HSV {
